@@ -46,9 +46,9 @@
 
     <properties>
         <!-- 项目版本 -->
-        <framework.components.revision>1.2.3.20220414-RELEASE</framework.components.revision>
+        <framework.components.revision>1.2.4.20231128-RELEASE</framework.components.revision>
 
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
 
         <!-- Deploy Properties -->
         <maven.deploy.skip>false</maven.deploy.skip>
@@ -74,193 +74,192 @@
         <jersey-client.version>1.19.4</jersey-client.version>
 
         <!-- dependencies  -->
-        <lombok.version>1.18.22</lombok.version>
-        <slf4j-api.version>1.7.36</slf4j-api.version>
-        <log4j.version>2.17.2</log4j.version>
-        <checker.version>3.21.4</checker.version>
+        <lombok.version>1.18.30</lombok.version>
+        <slf4j-api.version>2.0.9</slf4j-api.version>
+        <log4j.version>2.22.0</log4j.version>
+        <checker.version>3.40.0</checker.version>
 
         <!-- oshi -->
-        <oshi.version>6.1.6</oshi.version>
+        <oshi.version>6.4.8</oshi.version>
 
         <!-- dependencies management libs -->
         <!-- spring's version -->
-        <spring-cloud.version>2021.0.1</spring-cloud.version>
-        <spring-boot.version>2.6.6</spring-boot.version>
-        <spring-brick.version>3.0.1</spring-brick.version>
-        <spring-cloud-alibaba.version>2021.0.1.0</spring-cloud-alibaba.version>
+        <spring-cloud.version>2021.0.8</spring-cloud.version>
+        <spring-boot.version>2.7.18</spring-boot.version>
+        <spring-brick.version>3.1.2</spring-brick.version>
+        <spring-cloud-alibaba.version>2021.0.5.0</spring-cloud-alibaba.version>
         <!-- cache -->
-        <cache-caffeine.version>3.0.5</cache-caffeine.version>
+        <cache-caffeine.version>3.1.8</cache-caffeine.version>
         <!-- mybatis -->
-        <mybatis.version>3.5.9</mybatis.version>
-        <mybatis-spring.version>2.0.7</mybatis-spring.version>
+        <mybatis.version>3.5.14</mybatis.version>
+        <mybatis-spring.version>3.0.3</mybatis-spring.version>
         <mybatis-typehandlers-jsr310.version>1.0.2</mybatis-typehandlers-jsr310.version>
-        <mybatis-starter.version>2.2.2</mybatis-starter.version>
-        <mybatis-plus-starter.version>3.5.1</mybatis-plus-starter.version>
-        <mybatis-plus.version>3.5.1</mybatis-plus.version>
-        <mybatis-plus-generator.version>3.5.2</mybatis-plus-generator.version>
-        <mybatis-fluent.version>1.9.5</mybatis-fluent.version>
+        <mybatis-starter.version>3.0.3</mybatis-starter.version>
+        <mybatis-plus-starter.version>3.5.4.1</mybatis-plus-starter.version>
+        <mybatis-plus.version>3.5.4.1</mybatis-plus.version>
+        <mybatis-plus-generator.version>3.5.4.1</mybatis-plus-generator.version>
+        <mybatis-fluent.version>1.9.9</mybatis-fluent.version>
         <!-- cglib -->
         <cglib.version>3.3.0</cglib.version>
         <!-- disruptor -->
-        <disruptor.version>3.4.4</disruptor.version>
+        <disruptor.version>4.0.0</disruptor.version>
         <!-- pageHelper -->
-        <pagehelper.version>5.3.0</pagehelper.version>
-        <page-helper-starter.version>1.4.2</page-helper-starter.version>
+        <pagehelper.version>6.0.0</pagehelper.version>
+        <page-helper-starter.version>2.0.0</page-helper-starter.version>
         <!-- druid -->
-        <druid.version>1.2.9</druid.version>
+        <druid.version>1.2.20</druid.version>
         <!-- dozer -->
         <dozer.version>5.5.1</dozer.version>
         <!-- sky walking -->
-        <sky-walking.version>8.9.0</sky-walking.version>
-        <prometheus.varsion>1.8.4</prometheus.varsion>
+        <sky-walking.version>9.0.0</sky-walking.version>
+        <prometheus.varsion>1.12.0</prometheus.varsion>
         <!-- apache common -->
-        <apache-commons-lang3.version>3.12.0</apache-commons-lang3.version>
-        <apache-commons-net.version>3.8.0</apache-commons-net.version>
-        <apache-commons-io.version>2.11.0</apache-commons-io.version>
-        <apache-commons-compress.version>1.21</apache-commons-compress.version>
+        <apache-commons-lang3.version>3.14.0</apache-commons-lang3.version>
+        <apache-commons-net.version>3.10.0</apache-commons-net.version>
+        <apache-commons-io.version>2.15.0</apache-commons-io.version>
+        <apache-commons-compress.version>1.25.0</apache-commons-compress.version>
         <apache-commons-beanutils.version>1.9.4</apache-commons-beanutils.version>
-        <apache-commons-pool2.version>2.11.1</apache-commons-pool2.version>
-        <apache-commons-cli.version>1.4</apache-commons-cli.version>
-        <apache-commons-codec.version>1.15</apache-commons-codec.version>
+        <apache-commons-pool2.version>2.12.0</apache-commons-pool2.version>
+        <apache-commons-cli.version>1.6.0</apache-commons-cli.version>
+        <apache-commons-codec.version>1.16.0</apache-commons-codec.version>
         <apache-commons-configuration.version>1.10</apache-commons-configuration.version>
+        <apache-commons-configuration2.version>2.9.0</apache-commons-configuration2.version>
         <!-- reflections -->
         <reflections.version>0.10.2</reflections.version>
         <!-- hutool -->
-        <hutool.version>5.8.0.M2</hutool.version>
+        <hutool.version>5.8.23</hutool.version>
         <!-- concurrentlinkedhashmap-lru -->
         <concurrentlinkedhashmap-lru.version>1.4.2</concurrentlinkedhashmap-lru.version>
         <!-- json -->
-        <fastjson.version>1.2.80</fastjson.version>
-        <gson.version>2.9.0</gson.version>
-        <jackson.version>2.13.2</jackson.version>
+        <fastjson.version>2.0.42</fastjson.version>
+        <gson.version>2.10.1</gson.version>
+        <jackson.version>2.16.0</jackson.version>
         <!-- bouncy castle -->
         <bcprov-jdk15on.version>1.70</bcprov-jdk15on.version>
         <bcprov-jdk16.version>1.46</bcprov-jdk16.version>
         <bcprov-jdk15.version>1.46</bcprov-jdk15.version>
-        <bcprov-jdk15to18.version>1.70</bcprov-jdk15to18.version>
+        <bcprov-jdk15to18.version>1.77</bcprov-jdk15to18.version>
         <!-- swagger -->
         <swagger.version>3.0.0</swagger.version>
-        <swagger-annotations.version>2.1.13</swagger-annotations.version>
+        <swagger-annotations.version>2.2.19</swagger-annotations.version>
         <swagger-stater.version>2.0.2.RELEASE</swagger-stater.version>
         <!-- validation -->
-        <validation-api.version>2.0.1.Final</validation-api.version>
-        <hibernate-validator.version>6.2.3.Final</hibernate-validator.version>
-        <jakarta-validator.version>2.0.2</jakarta-validator.version>
+        <hibernate-validator.version>8.0.1.Final</hibernate-validator.version>
+        <jakarta-validator.version>3.0.2</jakarta-validator.version>
         <!-- dubbo -->
-        <dubbo.version>3.0.7</dubbo.version>
+        <dubbo.version>3.2.9</dubbo.version>
         <!-- zk client -->
-        <curator.version>5.2.1</curator.version>
+        <curator.version>5.5.0</curator.version>
         <zkclient.version>0.11</zkclient.version>
         <!-- xxl-job -->
-        <xxl-job.version>2.3.0</xxl-job.version>
+        <xxl-job.version>2.4.0</xxl-job.version>
         <!-- protostuff -->
         <io.protostuff.version>1.8.0</io.protostuff.version>
         <!-- kryo -->
-        <kryo.version>5.3.0</kryo.version>
+        <kryo.version>5.5.0</kryo.version>
         <!-- hessian -->
-        <hessian.version>4.0.65</hessian.version>
+        <hessian.version>4.0.66</hessian.version>
         <!-- msgpack -->
-        <msgpack.version>0.9.1</msgpack.version>
-        <msgpack-old.version>0.6.12</msgpack-old.version>
+        <msgpack.version>0.9.6</msgpack.version>
         <!-- db driver -->
-        <db-driver-mysql.version>8.0.28</db-driver-mysql.version>
-        <db-driver-h2.version>2.1.212</db-driver-h2.version>
+        <db-driver-mysql.version>8.2.0</db-driver-mysql.version>
+        <db-driver-h2.version>2.2.2242</db-driver-h2.version>
         <!-- redis -->
-        <lettuce.version>6.1.8.RELEASE</lettuce.version>
-        <jedis.version>4.2.2</jedis.version>
-        <redisson.version>3.17.0</redisson.version>
+        <lettuce.version>6.3.0.RELEASE</lettuce.version>
+        <jedis.version>5.1.0</jedis.version>
+        <redisson.version>3.24.3</redisson.version>
         <!-- mongodb -->
-        <mongo-driver.version>3.12.10</mongo-driver.version>
-        <mongo-driver-sync.version>4.5.1</mongo-driver-sync.version>
-        <morphia.version>1.3.2</morphia.version>
-        <morphia-core.version>2.2.6</morphia-core.version>
-        <morphia-core-old.version>1.6.1</morphia-core-old.version>
+        <mongo-driver.version>3.12.14</mongo-driver.version>
+        <mongo-driver-sync.version>4.11.1</mongo-driver-sync.version>
+        <morphia-core.version>2.4.8</morphia-core.version>
         <!-- lz4 -->
         <lz4.version>1.8.0</lz4.version>
         <!-- snappy -->
-        <snappy.version>1.1.8.4</snappy.version>
+        <snappy.version>1.1.10.5</snappy.version>
         <!-- guava -->
-        <google-guava.version>31.1-jre</google-guava.version>
-        <error_prone_annotations.version>2.12.1</error_prone_annotations.version>
+        <google-guava.version>32.1.3-jre</google-guava.version>
+        <error_prone_annotations.version>2.23.0</error_prone_annotations.version>
         <!-- servlet api-->
         <servlet-api.version>4.0.1</servlet-api.version>
         <!-- ali TTL -->
-        <ttl.version>2.12.6</ttl.version>
+        <ttl.version>2.14.3</ttl.version>
         <!-- 图片处理 -->
         <simpleImage.version>1.2.3</simpleImage.version>
         <jai.version>1.1.3</jai.version>
-        <thumbnailator.version>0.4.17</thumbnailator.version>
-        <twelvemonkeys.version>3.8.2</twelvemonkeys.version>
+        <thumbnailator.version>0.4.20</thumbnailator.version>
+        <twelvemonkeys.version>3.10.1</twelvemonkeys.version>
         <!-- 视频处理 -->
-        <media-analyzer.version>1.0.3</media-analyzer.version>
-        <javacv.version>1.5.7</javacv.version>
-        <opencv-platform.version>4.5.5-${javacv.version}</opencv-platform.version>
-        <ffmpeg-platform.version>5.0-${javacv.version}</ffmpeg-platform.version>
+        <media-analyzer.version>2.0.3</media-analyzer.version>
+        <javacv.version>1.5.9</javacv.version>
+        <javacv2.version>1.5.8</javacv2.version>
+        <opencv-platform.version>4.7.0-${javacv.version}</opencv-platform.version>
+        <ffmpeg-platform.version>6.0-${javacv.version}</ffmpeg-platform.version>
         <flycapture-platform.version>2.13.3.31-${javacv.version}</flycapture-platform.version>
-        <spinnaker-platform.version>2.4.0.143-${javacv.version}</spinnaker-platform.version>
+        <spinnaker-platform.version>3.0.0.118-${javacv.version}</spinnaker-platform.version>
         <libdc1394-platform.version>2.2.6-${javacv.version}</libdc1394-platform.version>
         <libfreenect-platform.version>0.5.7-${javacv.version}</libfreenect-platform.version>
         <libfreenect2-platform.version>0.2.0-${javacv.version}</libfreenect2-platform.version>
         <librealsense-platform.version>1.12.4-${javacv.version}</librealsense-platform.version>
-        <librealsense2-platform.version>2.50.0-${javacv.version}</librealsense2-platform.version>
+        <librealsense2-platform.version>2.53.1-${javacv.version}</librealsense2-platform.version>
         <videoinput-platform.version>0.200-${javacv.version}</videoinput-platform.version>
         <artoolkitplus-platform.version>2.3.1-${javacv.version}</artoolkitplus-platform.version>
-        <chilitags-platform.version>master-${javacv.version}</chilitags-platform.version>
-        <flandmark-platform.version>1.07-${javacv.version}</flandmark-platform.version>
-        <arrow-platform.version>6.0.1-${javacv.version}</arrow-platform.version>
-        <hdf5-platform.version>1.12.1-${javacv.version}</hdf5-platform.version>
-        <hyperscan-platform.version>5.4.0-${javacv.version}</hyperscan-platform.version>
-        <lz4-platform.version>1.9.3-${javacv.version}</lz4-platform.version>
-        <mkl-platform.version>2022.0-${javacv.version}</mkl-platform.version>
+        <chilitags-platform.version>master-${javacv2.version}</chilitags-platform.version>
+        <flandmark-platform.version>1.07-${javacv2.version}</flandmark-platform.version>
+        <arrow-platform.version>6.0.1-${javacv2.version}</arrow-platform.version>
+        <hdf5-platform.version>1.14.1-${javacv.version}</hdf5-platform.version>
+        <hyperscan-platform.version>5.4.2-${javacv.version}</hyperscan-platform.version>
+        <lz4-platform.version>1.9.4-${javacv2.version}</lz4-platform.version>
+        <mkl-platform.version>2023.1-${javacv.version}</mkl-platform.version>
         <mkl-dnn-platform.version>0.21.5-${javacv.version}</mkl-dnn-platform.version>
-        <dnnl-platform.version>2.5.2-${javacv.version}</dnnl-platform.version>
-        <openblas-platform.version>0.3.19-${javacv.version}</openblas-platform.version>
-        <arpack-ng-platform.version>3.8.0-${javacv.version}</arpack-ng-platform.version>
+        <dnnl-platform.version>3.1-${javacv.version}</dnnl-platform.version>
+        <openblas-platform.version>0.3.23-${javacv.version}</openblas-platform.version>
+        <arpack-ng-platform.version>3.9.0-${javacv.version}</arpack-ng-platform.version>
         <cminpack-platform.version>1.3.8-${javacv.version}</cminpack-platform.version>
         <fftw-platform.version>3.3.10-${javacv.version}</fftw-platform.version>
         <gsl-platform.version>2.7-${javacv.version}</gsl-platform.version>
-        <cpython-platform.version>3.10.2-${javacv.version}</cpython-platform.version>
-        <numpy-platform.version>1.22.2-${javacv.version}</numpy-platform.version>
-        <scipy-platform.version>1.8.0-${javacv.version}</scipy-platform.version>
-        <gym-platform.version>0.21.0-${javacv.version}</gym-platform.version>
-        <llvm-platform.version>13.0.1-${javacv.version}</llvm-platform.version>
-        <libffi-platform.version>3.4.2-${javacv.version}</libffi-platform.version>
-        <libpostal-platform.version>1.1-${javacv.version}</libpostal-platform.version>
-        <leptonica-platform.version>1.82.0-${javacv.version}</leptonica-platform.version>
-        <tesseract-platform.version>5.0.1-${javacv.version}</tesseract-platform.version>
-        <caffe-platform.version>1.0-${javacv.version}</caffe-platform.version>
-        <openpose-platform.version>1.7.0-${javacv.version}</openpose-platform.version>
-        <cuda-platform.version>11.6-8.3-${javacv.version}</cuda-platform.version>
-        <nvcodec-platform.version>11.1.5-${javacv.version}</nvcodec-platform.version>
+        <cpython-platform.version>3.11.3-${javacv.version}</cpython-platform.version>
+        <numpy-platform.version>1.24.3-${javacv.version}</numpy-platform.version>
+        <scipy-platform.version>1.10.0-${javacv.version}</scipy-platform.version>
+        <gym-platform.version>0.26.2-${javacv2.version}</gym-platform.version>
+        <llvm-platform.version>16.0.4-${javacv.version}</llvm-platform.version>
+        <libffi-platform.version>3.4.4-${javacv.version}</libffi-platform.version>
+        <libpostal-platform.version>1.1-${javacv2.version}</libpostal-platform.version>
+        <leptonica-platform.version>1.83.0-${javacv.version}</leptonica-platform.version>
+        <tesseract-platform.version>5.3.1-${javacv.version}</tesseract-platform.version>
+        <caffe-platform.version>1.0-${javacv2.version}</caffe-platform.version>
+        <openpose-platform.version>1.7.0-${javacv2.version}</openpose-platform.version>
+        <cuda-platform.version>12.1-8.9-${javacv.version}</cuda-platform.version>
+        <nvcodec-platform.version>12.1.14-${javacv.version}</nvcodec-platform.version>
         <opencl-platform.version>3.0-${javacv.version}</opencl-platform.version>
-        <mxnet-platform.version>1.9.0-${javacv.version}</mxnet-platform.version>
-        <pytorch-platform.version>1.10.2-${javacv.version}</pytorch-platform.version>
-        <tensorflow-platform.version>1.15.5-${javacv.version}</tensorflow-platform.version>
-        <tensorflow-lite-platform.version>2.8.0-${javacv.version}</tensorflow-lite-platform.version>
-        <tensorrt-platform.version>8.2-${javacv.version}</tensorrt-platform.version>
-        <ale-platform.version>0.7.3-${javacv.version}</ale-platform.version>
-        <depthai-platform.version>2.14.1-${javacv.version}</depthai-platform.version>
-        <onnx-platform.version>1.10.2-${javacv.version}</onnx-platform.version>
-        <ngraph-platform.version>0.26.0-${javacv.version}</ngraph-platform.version>
-        <onnxruntime-platform.version>1.10.0-${javacv.version}</onnxruntime-platform.version>
-        <tvm-platform.version>0.8.0-${javacv.version}</tvm-platform.version>
-        <liquidfun-platform.version>master-${javacv.version}</liquidfun-platform.version>
-        <qt-platform.version>5.15.2-${javacv.version}</qt-platform.version>
-        <skia-platform.version>2.80.3-${javacv.version}</skia-platform.version>
-        <cpu_features-platform.version>0.6.0-${javacv.version}</cpu_features-platform.version>
-        <modsecurity-platform.version>3.0.6-${javacv.version}</modsecurity-platform.version>
-        <opencv-platform-gpu.version>4.5.5-${javacv.version}</opencv-platform-gpu.version>
-        <caffe-platform-gpu.version>1.0-${javacv.version}</caffe-platform-gpu.version>
-        <mxnet-platform-gpu.version>1.9.0-${javacv.version}</mxnet-platform-gpu.version>
-        <pytorch-platform-gpu.version>1.10.2-${javacv.version}</pytorch-platform-gpu.version>
-        <tensorflow-platform-gpu.version>1.15.5-${javacv.version}</tensorflow-platform-gpu.version>
-        <onnxruntime-platform-gpu.version>1.10.0-${javacv.version}</onnxruntime-platform-gpu.version>
-        <tvm-platform-gpu.version>0.8.0-${javacv.version}</tvm-platform-gpu.version>
-        <tensorflow-platform-python.version>1.15.5-${javacv.version}</tensorflow-platform-python.version>
-        <tensorflow-platform-python-gpu.version>1.15.5-${javacv.version}</tensorflow-platform-python-gpu.version>
-        <cuda-platform-redist.version>11.6-8.3-${javacv.version}</cuda-platform-redist.version>
-        <tensorrt-platform-redist.version>8.2-${javacv.version}</tensorrt-platform-redist.version>
-        <mkl-platform-redist.version>2022.0-${javacv.version}</mkl-platform-redist.version>
+        <mxnet-platform.version>1.9.1-${javacv.version}</mxnet-platform.version>
+        <pytorch-platform.version>2.0.1-${javacv.version}</pytorch-platform.version>
+        <tensorflow-platform.version>1.15.5-${javacv2.version}</tensorflow-platform.version>
+        <tensorflow-lite-platform.version>2.12.0-${javacv.version}</tensorflow-lite-platform.version>
+        <tensorrt-platform.version>8.6-${javacv.version}</tensorrt-platform.version>
+        <tritonserver-platform.version>2.33-${javacv.version}</tritonserver-platform.version>
+        <ale-platform.version>0.8.0-${javacv2.version}</ale-platform.version>
+        <depthai-platform.version>2.21.2-${javacv.version}</depthai-platform.version>
+        <onnx-platform.version>1.14.0-${javacv.version}</onnx-platform.version>
+        <ngraph-platform.version>0.26.0-${javacv2.version}</ngraph-platform.version>
+        <onnxruntime-platform.version>1.15.0-${javacv.version}</onnxruntime-platform.version>
+        <tvm-platform.version>0.12.0-${javacv.version}</tvm-platform.version>
+        <liquidfun-platform.version>master-${javacv2.version}</liquidfun-platform.version>
+        <qt-platform.version>5.15.2-${javacv2.version}</qt-platform.version>
+        <skia-platform.version>2.88.3-${javacv2.version}</skia-platform.version>
+        <cpu_features-platform.version>0.7.0-${javacv2.version}</cpu_features-platform.version>
+        <modsecurity-platform.version>3.0.8-${javacv2.version}</modsecurity-platform.version>
+        <opencv-platform-gpu.version>4.7.0-${javacv.version}</opencv-platform-gpu.version>
+        <caffe-platform-gpu.version>1.0-${javacv2.version}</caffe-platform-gpu.version>
+        <mxnet-platform-gpu.version>1.9.1-${javacv2.version}</mxnet-platform-gpu.version>
+        <pytorch-platform-gpu.version>2.0.1-${javacv.version}</pytorch-platform-gpu.version>
+        <tensorflow-platform-gpu.version>1.15.5-${javacv2.version}</tensorflow-platform-gpu.version>
+        <onnxruntime-platform-gpu.version>1.15.0-${javacv.version}</onnxruntime-platform-gpu.version>
+        <tvm-platform-gpu.version>0.12.0-${javacv.version}</tvm-platform-gpu.version>
+        <tensorflow-platform-python.version>1.15.5-${javacv2.version}</tensorflow-platform-python.version>
+        <tensorflow-platform-python-gpu.version>1.15.5-${javacv2.version}</tensorflow-platform-python-gpu.version>
+        <cuda-platform-redist.version>12.1-8.9-${javacv.version}</cuda-platform-redist.version>
+        <tensorrt-platform-redist.version>8.6-${javacv.version}</tensorrt-platform-redist.version>
+        <mkl-platform-redist.version>2023.1-${javacv.version}</mkl-platform-redist.version>
         <!-- javax-el -->
         <javax-el.version>3.0.0</javax-el.version>
         <!-- glassfish-el -->
