@@ -4,9 +4,8 @@ import com.jphoebe.framework.components.util.constant.StringPoolConst;
 import com.jphoebe.framework.components.util.value.data.ObjectUtil;
 import com.jphoebe.framework.components.util.value.data.StringLengthUtil;
 import com.jphoebe.framework.components.util.value.validate.annotation.LengthSize;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 /**
  * @author 蒋时华
