@@ -1,13 +1,16 @@
 package com.jphoebe.framework.components.starter.mongodb.configuration;
 
-import com.jphoebe.framework.components.core.exception.defined.illegal.ParameterException;
 import com.jphoebe.framework.components.starter.mongodb.config.GlobalMongodbConfig;
 import com.jphoebe.framework.components.starter.mongodb.constant.MongodbConst;
 import com.jphoebe.framework.components.util.value.data.StrUtil;
 import com.mongodb.client.MongoClient;
 import dev.morphia.Datastore;
 import dev.morphia.Morphia;
-import dev.morphia.mapping.MapperOptions;
+import dev.morphia.config.MorphiaConfig;
+import io.smallrye.config.ConfigValuePropertiesConfigSource;
+import io.smallrye.config.SmallRyeConfigBuilder;
+import org.eclipse.microprofile.config.spi.ConfigSource;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.mongo.MongoProperties;
@@ -21,13 +24,17 @@ import org.springframework.data.mongodb.MongoTransactionManager;
 import org.springframework.data.mongodb.ReactiveMongoDatabaseFactory;
 import org.springframework.data.mongodb.ReactiveMongoTransactionManager;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+
 /**
  * @author 蒋时华
  * @date 2020-10-17 10:57:45
  */
 @Configuration
 @EnableConfigurationProperties(GlobalMongodbConfig.class)
-public class MorphiaConfig {
+public class MorphiaAutoConfiguration {
 
     @Bean
     public Datastore datastore(MongoClient mongoClient
@@ -51,16 +58,37 @@ public class MorphiaConfig {
         if (StrUtil.isBlank(databaseName)) {
             throw new RuntimeException("请配置mongodb相关参数");
         }
-
-        MapperOptions mapperOptions = MapperOptions.builder().mapSubPackages(true).build();
-        Datastore datastore = Morphia.createDatastore(mongoClient, databaseName, mapperOptions);
-        if (StrUtil.isBlank(globalMongodbConfig.getMapPackage())) {
-            throw new ParameterException("请填写，mongodb对应实体类包路径");
-        }
-        datastore.getMapper().mapPackage(globalMongodbConfig.getMapPackage());
-        // 确保实体类的映射建立
-        datastore.ensureIndexes();
-        return datastore;
+        HashMap<String, String> params = new HashMap<>();
+        params.put("morphia.apply-caps", globalMongodbConfig.getApplyCaps().toString());
+        params.put("morphia.apply-document-validations", globalMongodbConfig.getApplyDocumentValidations().toString());
+        params.put("morphia.apply-indexes", globalMongodbConfig.getApplyIndexes().toString());
+        params.put("morphia.auto-import-models", globalMongodbConfig.getAutoImportModels().toString());
+        params.put("morphia.codec-provider", globalMongodbConfig.getCodecProvider());
+        params.put("morphia.collection-naming", globalMongodbConfig.getCollectionNaming());
+        params.put("morphia.database", databaseName);
+        params.put("morphia.date-storage", globalMongodbConfig.getDateStorage());
+        params.put("morphia.discriminator", globalMongodbConfig.getDiscriminator());
+        params.put("morphia.discriminator-key", globalMongodbConfig.getDiscriminatorKey());
+        params.put("morphia.enable-polymorphic-queries", globalMongodbConfig.getEnablePolymorphicQueries().toString());
+        params.put("morphia.ignore-finals", globalMongodbConfig.getIgnoreFinals().toString());
+        params.put("morphia.packages", globalMongodbConfig.getMapPackage());
+        params.put("morphia.property-discovery", globalMongodbConfig.getPropertyDiscovery());
+        params.put("morphia.property-naming", globalMongodbConfig.getPropertyNaming());
+        params.put("morphia.query-factory", globalMongodbConfig.getQueryFactory());
+        params.put("morphia.store-empties", globalMongodbConfig.getStoreEmpties().toString());
+        params.put("morphia.store-nulls", globalMongodbConfig.getStoreNulls().toString());
+        params.put("morphia.uuid-representation", globalMongodbConfig.getUuidRepresentation());
+        ConfigValuePropertiesConfigSource configValue = new ConfigValuePropertiesConfigSource(params, "default", 0);
+        List<ConfigSource> configSources = new ArrayList<>();
+        configSources.add(configValue);
+        MorphiaConfig morphiaConfig = new SmallRyeConfigBuilder()
+                .addDefaultInterceptors()
+                .withMapping(MorphiaConfig.class)
+                .withSources(configSources)
+                .addDefaultSources()
+                .build()
+                .getConfigMapping(MorphiaConfig.class);
+        return Morphia.createDatastore(mongoClient, morphiaConfig);
     }
 
     @Bean(MongodbConst.TRANSACTION_NAME)
@@ -74,6 +102,7 @@ public class MorphiaConfig {
     @Bean(MongodbConst.TRANSACTION_REACTIVE_NAME)
     @ConditionalOnClass(name = "com.mongodb.reactivestreams.client.MongoClient")
     @ConditionalOnMissingBean(ReactiveMongoTransactionManager.class)
+    @ConditionalOnBean(ReactiveMongoDatabaseFactory.class)
     @Order(Ordered.LOWEST_PRECEDENCE)
     public ReactiveMongoTransactionManager transactionManager(ReactiveMongoDatabaseFactory factory) {
         return new ReactiveMongoTransactionManager(factory);
